# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie API",
    "version": "1.0.0",
    "description": "A RESTful API for managing movies, actors, and ratings."
  },
  "paths": {
...
    "/movies/{id}": {
      "get": {
        "summary": "Get a specific movie by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Movie details with HATEOAS links",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Movie" },
                    {
                      "type": "object",
                      "properties": {
                        "_links": {
                          "type": "object",
                          "properties": {
                            "self": { "type": "string" },
                            "ratings": { "type": "string" }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": { "description": "Movie not found" }
        }
      }
    },
    "/movies/{id}/ratings": {
      "get": {
        "summary": "Get ratings for a specific movie",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of ratings for a movie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "movie": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "title": { "type": "string" }
                      }
                    },
                    "ratings": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Rating" }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Movie not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "example": "The Matrix" },
          "release_year": { "type": "integer", "example": 1999 },
          "genre": { "type": "string", "example": "Sci-Fi" },
          "description": {
            "type": "string",
            "example": "A hacker discovers reality is a simulation."
          }
        },
        "required": ["title", "release_year", "genre"]
      },
      "Rating": {
        "type": "object",
        "properties": {
          "text": { "type": "string", "example": "Rating: 4.5" },
          "movie": {
            "type": "object",
            "properties": {
              "title": { "type": "string" }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}